# api/openapi.json
{
  "components": {
    "schemas": {
...
      "AppointmentOut": {
        "properties": {
          "appointment_id": {
            "format": "uuid",
            "title": "Appointment Id",
            "type": "string"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "inquiry_id": {
            "format": "uuid",
            "title": "Inquiry Id",
            "type": "string"
          },
          "replay": {
            "default": false,
            "title": "Replay",
            "type": "boolean"
          },
          "requested_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Text"
          },
          "scheduled_end": {
            "format": "date-time",
            "title": "Scheduled End",
            "type": "string"
          },
          "scheduled_start": {
            "format": "date-time",
            "title": "Scheduled Start",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "appointment_id",
          "inquiry_id",
          "status",
          "requested_text",
          "scheduled_start",
          "scheduled_end",
          "timezone",
          "external_ref"
        ],
        "title": "AppointmentOut",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceOut": {
        "properties": {
          "amount_paid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paid"
          },
          "counterparty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Name"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "discount_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Amount"
          },
          "discount_pct": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Pct"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "invoice_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Date"
          },
          "invoice_id": {
            "format": "uuid",
            "title": "Invoice Id",
            "type": "string"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          },
          "line_items": {
            "default": [],
            "items": {},
            "title": "Line Items",
            "type": "array"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "payment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Status"
          },
          "replay": {
            "default": false,
            "title": "Replay",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subtotal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtotal"
          },
          "tax_amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Amount"
          },
          "terms_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Hash"
          },
          "total": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "invoice_id",
          "invoice_number",
          "order_id",
          "direction",
          "status",
          "counterparty_name",
          "invoice_date",
          "due_date",
          "subtotal",
          "discount_pct",
          "discount_amount",
          "tax_amount",
          "total",
          "currency"
        ],
        "title": "InvoiceOut",
        "type": "object"
      },
...
      "OrderOut": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "discount_pct": {
            "title": "Discount Pct",
            "type": "string"
          },
          "inquiry_id": {
            "format": "uuid",
            "title": "Inquiry Id",
            "type": "string"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "quote_id": {
            "format": "uuid",
            "title": "Quote Id",
            "type": "string"
          },
          "replay": {
            "default": false,
            "title": "Replay",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "terms_hash": {
            "title": "Terms Hash",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "order_number",
          "quote_id",
          "inquiry_id",
          "status",
          "quantity",
          "discount_pct",
          "total",
          "currency",
          "terms_hash"
        ],
        "title": "OrderOut",
        "type": "object"
      },
...
      "QuoteOut": {
        "properties": {
          "approval_id": {
            "format": "uuid",
            "title": "Approval Id",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "discount_amount": {
            "title": "Discount Amount",
            "type": "string"
          },
          "discount_pct": {
            "title": "Discount Pct",
            "type": "string"
          },
          "inquiry_id": {
            "format": "uuid",
            "title": "Inquiry Id",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "quote_id": {
            "format": "uuid",
            "title": "Quote Id",
            "type": "string"
          },
          "quote_number": {
            "title": "Quote Number",
            "type": "string"
          },
          "replay": {
            "default": false,
            "title": "Replay",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subtotal": {
            "title": "Subtotal",
            "type": "string"
          },
          "terms_hash": {
            "title": "Terms Hash",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "string"
          },
          "unit_price": {
            "title": "Unit Price",
            "type": "string"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          }
        },
        "required": [
          "quote_id",
          "quote_number",
          "inquiry_id",
          "approval_id",
          "status",
          "quantity",
          "unit_price",
          "discount_pct",
          "subtotal",
          "discount_amount",
          "total",
          "currency",
          "terms_hash"
        ],
        "title": "QuoteOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lead-to-Cash PoC",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/inquiries/{inquiry_id}/appointment": {
      "post": {
        "operationId": "create_appointment_v1_inquiries__inquiry_id__appointment_post",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Calendar unavailable, retry later"
          }
        },
        "summary": "Create Appointment",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/v1/inquiries/{inquiry_id}/documents": {
      "get": {
        "description": "Evidence view: approval, quote, order and invoice side by side, with a consistency verdict.",
        "operationId": "documents_v1_inquiries__inquiry_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Documents V1 Inquiries  Inquiry Id  Documents Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/v1/inquiries/{inquiry_id}/quote": {
      "post": {
        "description": "Issue the quote from the granted approval (approved or modified terms, never the request).",
        "operationId": "create_quote_v1_inquiries__inquiry_id__quote_post",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Quote",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/inquiries/{inquiry_id}/trail": {
      "get": {
        "description": "Full evidence trail: request, every AI attempt, rule decision, approval, executed documents,\nexceptions with their resolution, and the ordered audit timeline.",
        "operationId": "trail_v1_inquiries__inquiry_id__trail_get",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Trail V1 Inquiries  Inquiry Id  Trail Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trail",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/v1/orders/{order_id}/invoice": {
      "post": {
        "description": "Generate the invoice as a DRAFT. It must pass validation (Step 7) before it can be approved.",
        "operationId": "create_invoice_v1_orders__order_id__invoice_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Invoice",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/v1/quotes/{quote_id}/order": {
      "post": {
        "description": "Customer accepted the quote: create the sales order with identical terms.",
        "operationId": "create_order_v1_quotes__quote_id__order_post",
        "parameters": [
          {
            "in": "path",
            "name": "quote_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Quote Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Order",
        "tags": [
          "documents"
        ]
      }
    }
  }
}